Type cast return 0 to unsigned int

+++ 0x07-pointers_arrays_strings/3-strspn.c
@@ -26,5 +26,5 @@ unsigned int _strspn(char *s, char *accept)
 		}
 	}
 
-	return (0);
+	return ((unsigned int)0);
 }
